test(todo_app_sqlite): rename directory
test(todo_app_sqlite): rename app directory
