+++ examples/todo_app_sqlite/e2e/tests/fixtures/checks.rs
@@ -1,4 +1,4 @@
-use crate::app::world::AppWorld;
+use super::world::AppWorld;
 use anyhow::Result;
 use fantoccini::Locator;
